+++ runall.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-# script to starup hsds docker containers
+# script to startup hsds docker containers
 if [ $# -eq 1 ] && ([ $1 == "-h" ] || [ $1 == "--help" ]); then
    echo "Usage: runall.sh [count]"
    exit 1
@@ -12,8 +12,10 @@ if [ $# -eq 1 ]; then
   count=$1
 fi
 
-echo "clean stopped containers"
-docker rm -v $(docker ps -aq -f status=exited) 
+if [ $(docker ps -aq -f status=exited | wc -l) -gt 0 ]; then
+   echo "clean stopped containers"
+   docker rm -v $(docker ps -aq -f status=exited) 
+fi
 
 echo "starting headnode"
 ./run.sh head $count
